+++ crates/slug-vm/tests/vm.rs
@@ -9,10 +9,10 @@ use std::{
 use slug_vm::SelectCase;
 use slug_vm::VmProgress;
 use slug_vm::{
-    CallArgumentKind, Capture, CaptureListId, Chunk, GlobalNameId, MatchMapKey, MatchPatternId,
-    MatchRest, ModuleLoader, NativeArity, NativeCall, NativeError, NativeModule, NativeOwnedValue,
-    NativeResourceType, NativeStatus, Op, Program, RuntimeErrorKind, SchemaField, SchemaFieldsId,
-    SourceSpan, SpanId, StructFieldsId, Value, Vm, compile,
+    CallArgumentKind, Capture, CaptureListId, Chunk, GlobalNameId, MatchMapKey, MatchPattern,
+    MatchPatternId, MatchRest, ModuleLoader, NativeArity, NativeCall, NativeError, NativeModule,
+    NativeOwnedValue, NativeResourceType, NativeStatus, Op, Program, RuntimeErrorKind, SchemaField,
+    SchemaFieldsId, SourceSpan, SpanId, StructFieldsId, Value, Vm, compile,
 };
 
 fn program_with_main(main: Chunk) -> Program {